* make sure "launch-threshold" key exists in availability array * split up & improve shortpixel copy, differentiating between launch & after launch * Uppercase the quality values in the dropdown

+++ classes/autoptimizeExtra.php
@@ -633,7 +633,7 @@ class autoptimizeExtra
             $avail_imgopt = $this->options['availabilities']['extra_imgopt'];
             $magic_number = intval( substr( md5( parse_url( AUTOPTIMIZE_WP_SITE_URL, PHP_URL_HOST ) ), 0, 3 ), 16 );
             $has_launched = get_option( 'autoptimize_imgopt_launched', '' );
-            if ( $has_launched || $magic_number < $avail_imgopt['launch-threshold'] ) {
+            if ( $has_launched || ( array_key_exists( 'launch-threshold', $avail_imgopt ) && $magic_number < $avail_imgopt['launch-threshold'] ) ) {
                 $launch_status = true;
                 if ( ! $has_launched ) {
                     update_option( 'autoptimize_imgopt_launched', 'on' );
@@ -752,7 +752,18 @@ class autoptimizeExtra
                 <th scope="row"><?php _e( 'Optimize Images', 'autoptimize' ); ?></th>
                 <td>
                     <label><input id='autoptimize_imgopt_checkbox' type='checkbox' name='autoptimize_extra_settings[autoptimize_extra_checkbox_field_5]' <?php if ( ! empty( $options['autoptimize_extra_checkbox_field_5'] ) && '1' === $options['autoptimize_extra_checkbox_field_5'] ) { echo 'checked="checked"'; } ?> value='1'><?php _e( 'Optimize images on the fly and serve them from a CDN.', 'autoptimize' ); ?></label>
-                    <?php echo apply_filters( 'autoptimize_extra_imgopt_settings_copy', '<p>' . __( 'Get more Google love and improve your website\'s loading speed by having the images optimized on the fly by', 'autoptimize' ) . ' <a href="https://shortpixel.com/aospai' . $sp_url_suffix . '" target="_blank">ShortPixel</a> ' . __( 'and then cached and served fast from a CDN. This is a free service up until a still to be defined threshold per domain, after which additional credits can be purchased.</p><p>Usage of this feature is subject to Shortpixel\'s', 'autoptimize' ) . '<a href="https://shortpixel.com/tos' . $sp_url_suffix . '" target="_blank">' . __( 'Terms of Use', 'autoptimize' ) . '</a> ' . __( 'and', 'autoptimize' ) . ' <a href="https://shortpixel.com/pp' . $sp_url_suffix . '" target="_blank">Privacy policy</a>.</p>' ); ?>
+                    <?php
+                    $upsell_msg_1 = '<p>' . __( 'Get more Google love and improve your website\'s loading speed by having the images optimized on the fly by', 'autoptimize' );
+                    $upsell_link  = ' <a href="https://shortpixel.com/aospai' . $sp_url_suffix . '" target="_blank">ShortPixel</a> ';
+                    $upsell_msg_2 = __( 'and then cached and served fast from a CDN.', 'autoptimize' ) . ' ';
+                    if ( 'launch' === $options['availabilities']['extra_imgopt']['status'] ) {
+                        $upsell_msg_3 = __( 'For a limited time only, this service is offered free-for-all, <b>don\'t miss the chance to test it</b> and see how much it could improve your site\'s speed.', 'autoptimize' );
+                    } else {
+                        $upsell_msg_3 = __( 'The service is offered for free for 100 images/month regardless of the traffic used. More image optimizations can be purchased starting with $4.99.', 'autoptimize' );
+                    }
+                    $terms_of_use = '</p><p>' . __( 'Usage of this feature is subject to Shortpixel\'s', 'autoptimize' ) . ' <a href="https://shortpixel.com/tos' . $sp_url_suffix . '" target="_blank">' . __( 'Terms of Use', 'autoptimize' ) . '</a> ' . __( 'and', 'autoptimize' ) . ' <a href="https://shortpixel.com/pp' . $sp_url_suffix . '" target="_blank">Privacy policy</a>.</p>';
+                    echo apply_filters( 'autoptimize_extra_imgopt_settings_copy', $upsell_msg_1 . $upsell_link . $upsell_msg_2 . $upsell_msg_3 . $terms_of_use );
+                    ?>
                 </td>
             </tr>
             <tr id='autoptimize_imgopt_quality' <?php if ( ! array_key_exists( 'autoptimize_extra_checkbox_field_5', $options ) || ( ! empty( $options['autoptimize_extra_checkbox_field_5'] ) && '1' !== $options['autoptimize_extra_checkbox_field_5'] ) ) { echo 'class="hidden"'; } ?>>
@@ -769,13 +780,13 @@ class autoptimizeExtra
                             if ( $_imgopt_val == $key ) {
                                 echo ' selected';
                             }
-                            echo '>' . $value . '</option>';
+                            echo '>' . ucfirst( $value ) . '</option>';
                         }
                         echo "\n";
                         ?>
                     </select>
                     </label>
-                    <p><?php echo apply_filters( 'autoptimize_extra_imgopt_quality_copy', '<a href="https://shortpixel.com/oic' . $sp_url_suffix .'" target="_blank">' . __( 'Test compression levels here.', 'autoptimize' ) . '</a>' ); ?></p>
+                    <p><?php echo apply_filters( 'autoptimize_extra_imgopt_quality_copy', __( 'You can', 'autoptimize' ) . ' <a href="https://shortpixel.com/oic' . $sp_url_suffix . '" target="_blank">' . __( 'test compression levels here', 'autoptimize' ) . '</a>.' ); ?></p>
                 </td>
             </tr>
             <tr>
